bug in srv/case_... (should use long for count/tot...): not fixed yet; move compute out of aggregationbuffer: not fixed yet

+++ ql/src/java/org/apache/hadoop/hive/ql/abm/udaf/CaseCountComputation.java
@@ -2,7 +2,7 @@ package org.apache.hadoop.hive.ql.abm.udaf;
 
 public class CaseCountComputation extends SrvCountComputation {
 
-  public void setCount(int cnt) {
+  public void setCount(long cnt) {
     this.baseCnt = cnt;
   }
